* Added `table.types.Value.CastTo(destination)` public method for cast `types.Value` to golang native type value destination * Dropped `table.types.CastTo` static helper
